+++ packages/react-libraries/CHANGELOG.md
@@ -15,6 +15,12 @@ All notable changes to this project will be documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Enhancements
+
+- The `doc` scripts warns about props defined in `defaultProps` but not in `propTypes`.
+
 ## 0.9.0 (July 16, 2018)
 
 ### Enhancements